Quick-win features: new datalinks, IPv6 ext headers, stats view, Wine test rig
- dissect: DLT_LINUX_SLL/SLL2 (Linux 'any' device) and DLT_NULL/DLT_LOOP
  (BSD/macOS loopback) framing — captures there previously showed only
  'Unknown datalink'; IPv6 extension-header chain is walked (hop-by-hop,
  routing, fragment, dest-options, AH) so such packets get real L4/L7
  dissection; non-first IPv4/IPv6 fragments are flagged instead of being
  misparsed as TCP/UDP garbage
- TUI: [V] protocol statistics overlay — uptime, totals, rates, drops,
  and a per-protocol breakdown with bars (counters existed, were never
  shown); stats_format no longer truncates on LLP64
- tests: regression cases for SLL/SLL2/NULL framing, ext-header walk,
  and both fragment guards (all with truncation sweeps)
- Windows testing: docker/windows-test + scripts/test-windows.sh cross-
  compile snuffles.exe and the unit suite with MinGW (static link) and
  run them under Wine; CI job replaces the compile-only cross check.
  Verified locally: all 4 suites pass under Wine.
- README: badges, TUI screenshot, updated feature/protocol tables

+++ src/stats.c
@@ -61,7 +61,7 @@ static void format_bytes(uint64_t bytes, char *buf, size_t len) {
     else if (bytes >= 1024ULL)
         snprintf(buf, len, "%.1f KB", (double)bytes / 1024.0);
     else
-        snprintf(buf, len, "%lu B", (unsigned long)bytes);
+        snprintf(buf, len, "%llu B", (unsigned long long)bytes);
 }
 
 static void format_rate(double bps, char *buf, size_t len) {
@@ -80,6 +80,6 @@ void stats_format(const stats_t *s, char *buf, size_t len) {
     format_bytes(s->total_bytes, bytes_str, sizeof(bytes_str));
     format_rate(s->bps, rate_str, sizeof(rate_str));
 
-    snprintf(buf, len, "Pkts: %lu  %s  PPS: %.1f  %s",
-             (unsigned long)s->total_packets, bytes_str, s->pps, rate_str);
+    snprintf(buf, len, "Pkts: %llu  %s  PPS: %.1f  %s",
+             (unsigned long long)s->total_packets, bytes_str, s->pps, rate_str);
 }
